refactor categories data flow to use raw categories + derived totals & move bootstrap trigger to ProfilePage
